fintech-api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fintech-api"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.5.2",
  "pydantic[email]>=2.0.0",
  "motor>=3.3.2",
  "pymongo>=4.6.1",
  "bcrypt>=4.0.0,<5.0.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "httpx>=0.25.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-mock>=3.11.0",
  "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["credentials*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
